create new utils module
