+++ herddb-core/src/test/java/herddb/utils/QueryUtilsTest.java
@@ -20,6 +20,8 @@
 package herddb.utils;
 
 import herddb.model.TableSpace;
+import herddb.utils.QueryUtils;
+import herddb.utils.QueryUtils;
 import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
