+++ herddb-utils/src/main/java/herddb/core/HerdDBInternalException.java
