+++ herddb-utils/src/main/java/herddb/utils/AbstractDataAccessor.java
