+++ herddb-utils/src/main/java/herddb/utils/AllNullsDataAccessor.java
