+++ herddb-utils/src/main/java/herddb/utils/BatchOrderedExecutor.java
