+++ herddb-utils/src/main/java/herddb/utils/BooleanHolder.java
