+++ herddb-utils/src/main/java/herddb/utils/ByteBufUtils.java
