+++ herddb-utils/src/main/java/herddb/utils/Bytes.java
