+++ herddb-utils/src/main/java/herddb/utils/ChangeThreadName.java
