+++ herddb-utils/src/main/java/herddb/utils/Constants.java
