+++ herddb-utils/src/main/java/herddb/utils/DefaultJVMHalt.java
