+++ herddb-utils/src/main/java/herddb/utils/EnsureIntegerIncrementAccumulator.java
