+++ herddb-utils/src/main/java/herddb/utils/EnsureLongIncrementAccumulator.java
