+++ herddb-utils/src/main/java/herddb/utils/ExtendedDataInputStream.java
