+++ herddb-utils/src/main/java/herddb/utils/ExtendedDataOutputStream.java
