+++ herddb-utils/src/main/java/herddb/utils/Holder.java
