+++ herddb-utils/src/main/java/herddb/utils/ManagedFile.java
