+++ herddb-utils/src/main/java/herddb/utils/MapUtils.java
