+++ herddb-utils/src/main/java/herddb/utils/MinDeltaLongIncrementAccumulator.java
