+++ herddb-utils/src/main/java/herddb/utils/NonClosingInputStream.java
