+++ herddb-utils/src/main/java/herddb/utils/NonClosingOutputStream.java
