+++ herddb-utils/src/main/java/herddb/utils/RawString.java
