+++ herddb-utils/src/main/java/herddb/utils/SQLUtils.java
