+++ herddb-utils/src/main/java/herddb/utils/SimpleByteArrayInputStream.java
