+++ herddb-utils/src/main/java/herddb/utils/SizeAwareObject.java
