+++ herddb-utils/src/main/java/herddb/utils/VisibleByteArrayOutputStream.java
