+++ herddb-utils/src/main/java/herddb/utils/Wrapper.java
