+++ herddb-utils/src/main/java/herddb/utils/XXHash64Utils.java
