+++ herddb-utils/src/test/java/herddb/utils/VisibleByteArrayOutputStreamTest.java
@@ -19,6 +19,8 @@
  */
 package herddb.utils;
 
+import herddb.utils.VisibleByteArrayOutputStream;
+import herddb.utils.XXHash64Utils;
 import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
 import java.util.Arrays;
